modernizing c types part E

+++ libs/gretchen/include/gretchen.h
@@ -33,9 +33,9 @@ typedef struct {
     size_t est_transfer_sec;
 } gretchenTX_inspect_t;
 
-void gretchenTX_inspect(gretchenTX_t* tx, uint8_t* filename, int* error, gretchenTX_inspect_t** info);
+void gretchenTX_inspect(gretchenTX_t* tx, uint8_t* filename, int8_t* error, gretchenTX_inspect_t** info);
 
-void gretchenTX_prepare(gretchenTX_t* tx, uint8_t* filename, int* error);
+void gretchenTX_prepare(gretchenTX_t* tx, uint8_t* filename, int8_t* error);
 
 void gretchenTX_get(gretchenTX_t* tx, float** samplebuffer, size_t* len);
 
@@ -58,7 +58,7 @@ gretchenRX_t* gretchenRX_create(grtModemOpt_t* opt, int internal_bufsize);
 
 void gretchenRX_destroy(gretchenRX_t* rx);
 
-void gretchenRX_push_le16f(gretchenRX_t* rx, float* buffer, size_t len, int* error);
+void gretchenRX_push_le16f(gretchenRX_t* rx, float* buffer, size_t len, int8_t* error);
